Slight refactor of ListUnspentResponse, and some wish_list cleanups
- ListUnspent has been slightly refactored to be a little bit more
efficient by allowing for the same script_hash appearing in the same tx
multiple times to be efficiently encoded, rather than reproducing the
txid each time on the wire when this occurs.

- Added the TXOInfoRespone oneof field to indicate a requested txo info
was not found.  This is preferable to returning a full on error if just
1 txo is missing.  It is helpful to know which txo is invalid in a list
for plugins like CashShuffle, CashFusion, etc.