Refactored: Transaction merkle info into its own message, and more
- Transaction service has a new message, TxMerkleInfo, which is re-used
by both the TransactionResponse and the MerkleResponse messages to
contain merkle information for a transaction. In this way the messages
themselves can be serialized/cached in the db, etc.

- Script_hash.History has a more complete pagination scheme.  Added a
"ContinuationSpec" message which is returned by the server to tell the
client how to continue downloading the history in cases only where the
result set couldn't be retrieved in one go (eg because it exceeded
server limits on how much history a client can retrieve at once). This
effectively allows for proper pagination of histories for large wallets.

RFC.